Removed redundant tests.
