+++ src/Tests/Pure.Primitives.Choices.Tests/DayOfWeek/DayOfWeekChoiceTests.cs
@@ -1,4 +1,5 @@
-﻿using Pure.Primitives.Bool;
+﻿using Pure.Primitives.Abstractions.DayOfWeek;
+using Pure.Primitives.Bool;
 using Pure.Primitives.DayOfWeek;
 
 namespace Pure.Primitives.Tests.DayOfWeek;
