+++ src/Tests/Pure.Primitives.Choices.Tests/Pure.Primitives.Choices.Tests.csproj
@@ -10,8 +10,13 @@
   <ItemGroup>
     <PackageReference Include="coverlet.collector" Version="6.0.2" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.12.0" />
+    <PackageReference Include="Pure.Primitives" Version="1.0.0" />
     <PackageReference Include="xunit" Version="2.9.2" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.8.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Pure.Primitives.Choices\Pure.Primitives.Choices.csproj" />
   </ItemGroup>
 
   <ItemGroup>
